+++ verl/utils/kernel/linear_cross_entropy.py
@@ -39,7 +39,7 @@ from . import kernels
 
 class LinearCrossEntropy(torch.autograd.Function):
     @staticmethod
-    def forward(ctx, hidden: torch.Tensor, weight: torch.Tensor, labels: torch.Tensor, reduction: typing.Optional[str] = "mean", temperature: typing.Optional[float] = 1.0, dist_process_group: typing.Optional[dist.ProcessGroup] = None) -> typing.List[torch.Tensor]:
+    def forward(ctx, hidden: torch.Tensor, weight: torch.Tensor, labels: torch.Tensor, reduction: typing.Optional[str] = "none", temperature: typing.Optional[float] = 1.0, dist_process_group: typing.Optional[dist.ProcessGroup] = None) -> typing.List[torch.Tensor]:
         """_summary_
 
         Args:
@@ -47,7 +47,7 @@ class LinearCrossEntropy(torch.autograd.Function):
             hidden (torch.Tensor): (batch_size, num_tokens, hidden_size) -> (batch_size * num_tokens, hidden_size)
             weight (torch.Tensor): (vocab_size, hidden_size)
             labels (torch.Tensor): (batch_size, num_tokens) -> (batch_size * num_tokens, )
-            reduction (typing.Optional[str], optional): _description_. Defaults to "mean".
+            reduction (typing.Optional[str], optional): _description_. Defaults to "none".
             temperature (typing.Optional[float], optional): _description_. Defaults to 1.0.
             dist_process_group (typing.Optional[dist.ProcessGroup], optional): _description_. Defaults to None.
 
@@ -58,6 +58,7 @@ class LinearCrossEntropy(torch.autograd.Function):
         with torch.cuda.nvtx.range("LinearCrossEntropy-forward"):
             REDUCTION = kernels.get_entropy_reduction_enum_number(reduction.lower())
 
+            original_hidden_shape = hidden.shape
             if len(hidden.shape) != 2:
                 hidden = hidden.view(-1, hidden.shape[-1])  # (batch_size * num_tokens, hidden_size)
             if len(labels.shape) != 1:
@@ -66,6 +67,7 @@ class LinearCrossEntropy(torch.autograd.Function):
             logprobs, entropy, _maximum, _accumulate, _entropy_b = kernels.efficient_entropy_forward(hidden, weight, labels, REDUCTION, temperature, dist_process_group)
 
             ctx.save_for_backward(hidden, weight, labels, _maximum, _accumulate, _entropy_b)
+            ctx.original_hidden_shape = original_hidden_shape
             ctx.REDUCTION = REDUCTION
             ctx.dist_process_group = dist_process_group
             ctx.should_return_fp32_grad = False
@@ -82,89 +84,9 @@ class LinearCrossEntropy(torch.autograd.Function):
             temperature = ctx.temperature
 
             d_hidden, d_weight = kernels.efficient_entropy_backward(dlogprobs, dentropy, hidden, weight, labels, _maximum, _accumulate, _entropy_b, REDUCTION, should_return_fp32_grad, temperature, dist_process_group)
+            d_hidden = d_hidden.view(ctx.original_hidden_shape)
 
         return (d_hidden, d_weight, None, None, None, None)
 
 
 linear_cross_entropy = LinearCrossEntropy.apply
-
-
-class TorchEntropyTP(torch.autograd.Function):
-    """
-    it is used for testing the correctness of the kernel
-    it is not efficient and is not recommended to use in practice
-    """
-
-    @staticmethod
-    def forward(ctx, hidden: torch.Tensor, weight: torch.Tensor, labels: torch.Tensor, temperature: float, dist_process_group: torch.distributed.ProcessGroup):
-        # weight has shape [vocab_size, hidden_size], hidden has shape [num_tokens, hidden_size]
-        logits = torch.matmul(hidden.to(torch.float32), weight.to(torch.float32).T)  # [num_tokens, vocab_size]
-        logits /= temperature
-        whole_logits = torch.empty((logits.shape[0], logits.shape[1] * dist.get_world_size(dist_process_group)), dtype=logits.dtype, device=logits.device)
-        whole_logits_ref = [whole_logits[:, i * logits.shape[1] : (i + 1) * logits.shape[1]] for i in range(dist.get_world_size(dist_process_group))]
-        dist.all_gather(whole_logits_ref, logits, group=dist_process_group)
-
-        pd = torch.nn.functional.softmax(whole_logits, dim=-1)
-        entropy_a = torch.logsumexp(whole_logits, dim=-1)  # [num_tokens]
-        entropy_b = torch.sum(pd * whole_logits, dim=-1)  # [num_tokens]
-        entropy = entropy_a - entropy_b
-
-        logprobs = torch.nn.functional.cross_entropy(whole_logits, labels, reduction="none")
-        logprobs = torch.neg(logprobs)
-
-        ctx.save_for_backward(hidden, weight, labels, whole_logits, entropy_b)
-        ctx.dist_process_group = dist_process_group
-        ctx.temperature = temperature
-        return logprobs, entropy
-
-    @staticmethod
-    def backward(ctx, g_logprobs: torch.Tensor, g_entropy: torch.Tensor):
-        hidden, weight, labels, whole_logits, entropy_b = ctx.saved_tensors
-        dist_process_group = ctx.dist_process_group
-        temperature = ctx.temperature
-        batch_size, hidden_size = hidden.shape
-        vocab_size, hidden_size = weight.shape
-        rank = dist.get_rank(dist_process_group)
-
-        # Compute softmax probabilities
-        maximum, _ = torch.max(whole_logits, dim=-1, keepdim=True)
-        exp_logits = torch.exp(whole_logits - maximum)
-        accumulate = exp_logits.sum(dim=-1, keepdim=True)
-        pd = exp_logits / accumulate
-
-        # Gradient for entropy
-        # entropy = entropy_a - entropy_b
-        # entropy_a = log(sum(exp(logits)))
-        # entropy_b = sum(pd * logits)
-        # d_entropy_a/d_logits = pd
-        # d_entropy_b/d_logits = pd * (logits - b.unsqueeze(1) + 1)
-        # d_entropy/d_logits = d_entropy_a - d_entropy_b
-        # d_entropy/d_logits = pd - pd * (logits - b.unsqueeze(1) + 1)
-        # d_entropy/d_logits = -pd * (logits - b.unsqueeze(1))
-        d_logits_entropy = g_entropy.unsqueeze(1) * (-pd * (whole_logits - entropy_b.unsqueeze(1)))
-
-        # Gradient for logprobs
-        # logprobs = -cross_entropy = -log(pd[labels])
-        # d_logprobs/d_logits = (pd - one_hot(labels))
-        one_hot = torch.zeros_like(whole_logits)
-        one_hot.scatter_(1, labels.unsqueeze(1), 1)
-        g_logprobs = torch.neg(g_logprobs)
-        d_logits_logprobs = g_logprobs.unsqueeze(1) * (pd - one_hot)
-        # NOTE: This will lead to wrong result
-        # d_logits_logprobs = g_logprobs.unsqueeze(1) * (pd - 1) * one_hot
-
-        # Combine gradients
-        d_logits = d_logits_entropy + d_logits_logprobs
-        d_logits /= temperature
-
-        # Get local slice of gradients
-        local_d_logits = d_logits[:, rank * vocab_size : (rank + 1) * vocab_size]
-
-        # Compute gradients for hidden and weight
-        d_hidden = torch.matmul(local_d_logits, weight.to(torch.float32))
-        d_weight = torch.matmul(local_d_logits.T, hidden.to(torch.float32))
-
-        return d_hidden, d_weight, None, None, None
-
-
-run_torch_entropy_tp = TorchEntropyTP.apply
